+Updated Assignmenet 9

+++ Assignment 9/Code/Features.java
@@ -1,12 +1,12 @@
 abstract class Features implements IFeatures{
 
-	public Features(int size, int xPosition, int yPosition){
+	public Features(float size, float xPosition, float yPosition){
 		this.size = size;
 		this.xPosition = xPosition;
 		this.yPosition = yPosition;
 	}
 	
 
-	public int size;
-	public int xPosition, yPosition;
+	public float size;
+	public float xPosition, yPosition;
 }
